Updating WhatsAppWeb client version
It seems there is New minor update to 2035.
and I note that WhatsApp changes the client version to use the OS version. 
Well, I take back the update as the @tulir Advice. Thanks, tulir  : https://github.com/Rhymen/go-whatsapp/pull/431#issuecomment-681146713
But I kept function names as it is so there is no backward compatibility problem